thrift: move TestSkipUUIDField into decode_test.go
